feat: re-implement functionality using serverless functions
Resolves #11.
